% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_nat_inf.R
\name{do_sens_mono_nat_inf}
\alias{do_sens_mono_nat_inf}
\title{Sensitivity analysis to monotonicity using AIPW estimator}
\usage{
do_sens_mono_nat_inf(
  data,
  models,
  Y_name = "Y",
  Z_name = "Z",
  S_name = "S",
  epsilon = exp(seq(log(0.5), log(2), length = 49))
)
}
\arguments{
\item{data}{A data.frame containing observed data.}

\item{models}{A named list of fitted nuisance models.}

\item{Y_name}{Character string specifying the outcome variable.}

\item{Z_name}{Character string specifying the treatment variable.}

\item{S_name}{Character string specifying the infection indicator.}

\item{epsilon}{Numeric vector of sensitivity parameters.}
}
\value{
A list with class \code{"monotonicity_sens"} containing:
\describe{

}
}
\description{
Performs sensitivity analysis for violations of monotonicity using
a parameter \eqn{\epsilon}.
}
