% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{provide}
\alias{provide}
\title{Example dataset inspired by the PROVIDE study}
\format{
A data frame with multiple observations and the following variables:
\describe{
\item{wk10_haz}{Height-for-age Z-score at 10 weeks (numeric).}
\item{gender}{Infant gender ("Male", "Female").}
\item{num_hh_sleep}{Number of individuals sleeping in the household (integer).}
\item{rotaarm}{Vaccination indicator (0 = control, 1 = vaccine).}
\item{rotaepi}{Infection indicator (0 = no episode, 1 = episode).}
\item{any_abx_wk52}{Any antibiotic use by 52 weeks (0/1).}
}
}
\source{
Simulated data loosely based on the PROVIDE study.
}
\usage{
provide
}
\description{
A simulated dataset based on PROVIDE study data for analysis of vaccine on antibiotic use
}
\keyword{datasets}
